#10 カレンダーの関連の顧客担当者が編集時に消える問題を修正

+++ modules/Calendar/Activity.php
@@ -173,24 +173,27 @@ class Activity extends CRMEntity {
 		}
 
 		$recordId = intval($this->id);
-		if(isset($_REQUEST['contactidlist']) && $_REQUEST['contactidlist'] != '') {
-			$adb->pquery( 'DELETE from vtiger_cntactivityrel WHERE activityid = ?', array($recordId));
 
-			$contactIdsList = explode (';', $_REQUEST['contactidlist']);
-			$count = count($contactIdsList);
-            $params=array();
-			$sql = 'INSERT INTO vtiger_cntactivityrel VALUES ';
-			for($i=0; $i<$count; $i++) {
-				$contactIdsList[$i] = intval($contactIdsList[$i]);
-				$sql .= " (?, ?)";
-				array_push($params,$contactIdsList[$i],$recordId);
-				if ($i != $count - 1) {
-					$sql .= ',';
+		//ドラッグ&ドロップ時にはContactsは変更されずcontactidlistを得られないので関連を更新しない
+		if(empty($_REQUEST['isDragDrop'])){
+			if(isset($_REQUEST['contactidlist']) && $_REQUEST['contactidlist'] != '') {
+				$adb->pquery( 'DELETE from vtiger_cntactivityrel WHERE activityid = ?', array($recordId));
+				$contactIdsList = explode (';', $_REQUEST['contactidlist']);
+				$count = count($contactIdsList);
+				$params=array();
+				$sql = 'INSERT INTO vtiger_cntactivityrel VALUES ';
+				for($i=0; $i<$count; $i++) {
+					$contactIdsList[$i] = intval($contactIdsList[$i]);
+					$sql .= " (?, ?)";
+					array_push($params,$contactIdsList[$i],$recordId);
+					if ($i != $count - 1) {
+						$sql .= ',';
+					}
 				}
+				$adb->pquery($sql, $params);
+			} else if ($_REQUEST['contactidlist'] == '' && $insertion_mode == "edit") {
+				$adb->pquery('DELETE FROM vtiger_cntactivityrel WHERE activityid = ?', array($recordId));
 			}
-			$adb->pquery($sql, $params);
-		} else if ($_REQUEST['contactidlist'] == '' && $insertion_mode == "edit") {
-			$adb->pquery('DELETE FROM vtiger_cntactivityrel WHERE activityid = ?', array($recordId));
 		}
 
 		//Insert into cntactivity rel
